+++ src/index.js
@@ -2,6 +2,7 @@
 // book. Import a subpath (`book-of-elementals/accordion`) to register just one.
 export * from './core.js';
 export * from './elementals/accordion/index.js';
+export * from './elementals/combobox/index.js';
 export * from './elementals/disclosure/index.js';
 export * from './elementals/menu/index.js';
 export * from './elementals/navbar/index.js';